requirements.txt
openai
python-dotenv
sentence-transformers
psycopg2-binary
requests
pillow
pgvector
streamlit
#Scrapy
#itemadapter